+++ src/Map/src/Bridge/Google/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 -   Minimum required Symfony version is now 6.4
 -   Minimum required PHP version is now 8.2
+-   Upgrade `@googlemaps/js-api-loader` JavaScript package from `^1.16.6` to `^2.0.0`:
+    - If you use Symfony AssetMapper without Symfony Flex, run `bin/console importmap:require @googlemaps/js-api-loader@^2.0`
+    - Options configurable through `UX_MAP_DSN` query params have changed:
+        - Option `version` has been renamed to `v`
+        - Options `authReferrerPolicy`, `mapIds`, `channel`, `solutionChannel` have been added
+        - Options `ìd`, `nonce`, `retries`, `url` have been removed
 
 ## 2.31